+++ CHANGELOG.md
@@ -3,9 +3,30 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [7.10.9] - 2025-05-22
+## [7.10.14] - 2025-05-22
 ### Changed
 - Optional variable to disable glue database prefix.
+
+## [7.10.13] - 2025-05-19
+### Added
+- Backwards compatibility with `var.lf_catalog_glue_sync_arn`.
+
+## [7.10.12] - 2025-05-19
+### Added
+- Changed from LF permissions `aws_iam_role.apiary_hms_readwrite.arn` to `var.lf_catalog_glue_sync_arn`.
+
+## [7.10.11] - 2025-05-16
+### Added
+- Added variables to support adding custom log4j configuration for readonly,readwrite and housekeeper hms when deployed in Kubernetes.
+
+## [7.10.10] - 2025-05-14
+### Added
+- Added variable `s3_logs_customer_accounts` to support allowing read access on s3 logs bucket.
+- Added varialbe `s3_inventory_expiration_days` to support s3 inventory bucket lifecycle.
+
+## [7.10.9] - 2025-05-14
+### Added
+- Added outputs for apiary data buckets arns and apiary system bucket arn
 
 ## [7.10.8] - 2025-05-12
 ### Changed